requirements.txt
discord.py
python-dotenv
pillow
pyyaml
flask
psycopg2-binary